App completed. Built ViewEmployees in HomeController.
1:04:26 - FINISH

This section built out ViewEmployees() on HomeController, which moved the data from DB to frontend at last. I auto-created the View for ViewEmployees(); steps for that process are in the code comments above that function.

I'm going to map this out on my Miro board and describe the ViewEmployees() in my notebook. My next stage is to map out the RCO QA project in a similar way so I can understand it better. Future Me: I hope that this repo is as resourceful as I tried to make it this week.